requirements.txt
requests
pymsteams
pandas
pyspark
python-decouple
azure-storage-blob
Flask

